requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary
redis==5.0.1
celery==5.3.4
stripe==7.8.0
paypalrestsdk==1.13.3
sendgrid==6.10.0
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
django-environ==0.11.2
django-storages==1.14.2
boto3==1.34.0
elasticsearch==8.11.0
django-elasticsearch-dsl==8.0.0
django-redis==5.4.0
django-cacheops==7.0.1
django-ratelimit==4.1.0
django-import-export==3.3.1
django-cleanup==9.0.0
django-compressor==4.4
django-heroku==0.3.1
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
